--- SpringHiberStarter/src/main/java/org/itstep/service/LessonServiceImpl.java
@@ -1,8 +1,0 @@
-package org.itstep.service;
-
-public class LessonServiceImpl implements LessonService {
-
-
-
-
-}
