+++ SpringHiberStarter/src/main/java/org/itstep/service/implementation/LessonServiceImpl.java
@@ -1,0 +1,10 @@
+package org.itstep.service.implementation;
+
+import org.itstep.service.LessonService;
+
+public class LessonServiceImpl implements LessonService {
+
+
+
+
+}
