+++ SpringHiberStarter/src/main/java/org/itstep/service/implementation/StudentServiceImpl.java
@@ -1,8 +1,9 @@
-package org.itstep.service;
+package org.itstep.service.implementation;
 
 
 import org.itstep.dao.StudentDAO;
 import org.itstep.dao.pojo.Student;
+import org.itstep.service.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
